% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{layout_islands}
\title{Deprecated functions}
\usage{
layout_islands(plot, ryukyu = TRUE, ogasawara = TRUE)
}
\arguments{
\item{plot}{A plot of ggplot2.}

\item{ryukyu}{Include the Ryukyu Islands in the map? By default, `TRUE`.}

\item{ogasawara}{Include the Ogasawara Islands in the map? By default,
`TRUE`.}
}
\description{
Deprecated functions
}
