% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_japan.R
\name{layout_japan}
\alias{layout_japan}
\title{Lay out the Japanese prefectures}
\usage{
layout_japan(plot, ryukyu = TRUE, ogasawara = TRUE)
}
\arguments{
\item{plot}{A plot of ggplot2.}

\item{ryukyu}{Include the Ryukyu Islands in the map? By default, `TRUE`.}

\item{ogasawara}{Include the Ogasawara Islands in the map? By default,
`TRUE`.}
}
\value{
A plot of ggplot2.
}
\description{
Lay out the Japanese prefectures
}
