% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prefecture}
\alias{prefecture}
\title{Japanese prefectures}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 47 rows and 4 columns.
}
\source{
<https://CRAN.R-project.org/package=mapdata>
}
\usage{
prefecture
}
\description{
Japanese prefectures
}
\keyword{datasets}
